More cleanup

+++ gamestate-traits/src/lib.rs
@@ -109,8 +109,8 @@ pub trait MachinationTrait {
                     let colour = self.get_palette().0[*p];
                     for _ in 0..f {
                         pixels.set_pixel(
-                            (x + xtmp as i32 - n as i32) as usize, // - (image.left_offset as i32),
-                            (y + ytmp as i32 + c.y_offset as i32 * f as i32) as usize, // - image.top_offset as i32 - 30,
+                            (x + xtmp - n as i32) as usize, // - (image.left_offset as i32),
+                            (y + ytmp + c.y_offset * f as i32) as usize, // - image.top_offset as i32 - 30,
                             colour.r,
                             colour.g,
                             colour.b,
